# openapi.yaml
openapi: 3.0.3
info:
  title: distrebute.com Platform API
  version: 0.1.0
  description: |
    Unified API surface for distrebute.com — the 22 UI templates wired to
    backend services. Every endpoint is owned by exactly one microservice
    (see `x-owner` extension). Gateway routes path prefixes to services.

    Auth model: short-lived JWT access tokens (15 min) + opaque refresh
    tokens via the auth-service. Passkey / WebAuthn primary, password fallback.
  contact:
    name: distrebute.com platform team
    email: api@distrebute.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local docker-compose gateway
  - url: https://api.distrebute.com/v1
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /users/me/profiles:
    get:
      tags: [users]
      summary: List family profiles (adult + kid)
      x-owner: user-service
      x-templates: [21]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Profile' }
    post:
      tags: [users]
      summary: Create a new profile
      x-owner: user-service
      x-templates: [21]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
  /users/me/profiles/{profileId}/parental:
    put:
      tags: [users]
      summary: Update parental controls (age cap, screen time, schedule, categories)
      x-owner: user-service
      x-templates: [21]
      parameters:
        - in: path
          name: profileId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ParentalControls' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParentalControls' }
  /users/me/pin:
    post:
      tags: [users]
      summary: Set household PIN to exit kid mode
      x-owner: user-service
      x-templates: [21]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pin: { type: string, minLength: 4, maxLength: 6 }
      responses:
        '204': { description: PIN set }
  /users/me/privacy:
    get:
      tags: [users]
      summary: Privacy toggle states (history, personalization, etc.)
      x-owner: user-service
      x-templates: [19]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrivacySettings' }
    put:
      tags: [users]
      summary: Update privacy toggles
      x-owner: user-service
      x-templates: [19]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PrivacySettings' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrivacySettings' }
...
  /users/me/devices/{deviceId}:
    delete:
      tags: [users]
      summary: Sign out a specific device
      x-owner: user-service
      x-templates: [20]
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string }
      responses:
        '204': { description: Signed out }
...
  /users/me/devices/{deviceId}/not-me:
    post:
      tags: [users]
      summary: Mark a sign-in attempt as fraudulent, trigger recovery flow
      x-owner: user-service
      x-templates: [20]
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string }
      responses:
        '202': { description: Recovery flow initiated }

  # ---- METADATA: titles, episodes, channels --------------------------------
...
  /social/follow/{channelId}:
    post:
      tags: [social]
      summary: Subscribe to a creator
      x-owner: social-service
      x-templates: [4]
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string }
      responses:
        '204': { description: subscribed }
    delete:
      tags: [social]
      summary: Unsubscribe from a creator
      x-owner: social-service
      x-templates: [4]
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string }
      responses:
        '204': { description: unsubscribed }
  /social/comments/{videoId}:
    get:
      tags: [social]
      summary: List comments on a video
      x-owner: social-service
      x-templates: [2]
      parameters:
        - in: path
          name: videoId
          required: true
          schema: { type: string }
        - { in: query, name: sort, schema: { type: string, enum: [top, new] } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Comment' }
    post:
      tags: [social]
      summary: Post a comment
      x-owner: social-service
      x-templates: [2]
      parameters:
        - in: path
          name: videoId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body: { type: string }
                parentId: { type: string, nullable: true }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
  /social/watch-party:
    post:
      tags: [social]
      summary: Create a new watch party room
      x-owner: social-service
      x-templates: [12]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                videoId: { type: string }
                inviteOnly: { type: boolean }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchParty' }
  /social/watch-party/{roomId}/join:
    post:
      tags: [social]
      summary: Join a watch party via invite link
      x-owner: social-service
      x-templates: [12]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  wsUrl: { type: string }
                  participants:
                    type: array
                    items: { $ref: '#/components/schemas/Participant' }
  /social/watch-party/{roomId}/react:
    post:
      tags: [social]
      summary: Send a floating emoji reaction to the room
      x-owner: social-service
      x-templates: [12]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                emoji: { type: string }
      responses:
        '204': { description: broadcast }

  # ---- NOTIFICATIONS -------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Profile:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        avatarUrl: { type: string }
        age: { type: integer, nullable: true }
        isKidProfile: { type: boolean }
        ageRatingCap: { type: string, enum: [U, U/A 7+, U/A 13+, U/A 16+, A] }
    ProfileCreate:
      type: object
      required: [name, isKidProfile]
      properties:
        name: { type: string }
        avatarUrl: { type: string }
        age: { type: integer }
        isKidProfile: { type: boolean }
        ageRatingCap: { type: string }
    ParentalControls:
      type: object
      properties:
        ageRatingCap: { type: string }
        dailyScreenTimeMinutes: { type: integer }
        weeklySchedule:
          type: array
          description: 7 entries (Mon→Sun), each with allowed time ranges
          items:
            type: object
            properties:
              day: { type: integer, minimum: 0, maximum: 6 }
              allowedRanges:
                type: array
                items:
                  type: object
                  properties:
                    startMinute: { type: integer }
                    endMinute: { type: integer }
        allowedCategories:
          type: array
          items: { type: string }
        blockedCategories:
          type: array
          items: { type: string }
        pinRequiredToExit: { type: boolean }
    PrivacySettings:
      type: object
      properties:
        watchHistory: { type: boolean }
        personalization: { type: boolean }
        voiceSearchHistory: { type: boolean }
        adPersonalization: { type: boolean }
        locationForLive: { type: boolean }
        crashDiagnostics: { type: boolean }
...
    Comment:
      type: object
      properties:
        id: { type: string }
        videoId: { type: string }
        parentId: { type: string, nullable: true }
        body: { type: string }
        author:
          type: object
          properties:
            handle: { type: string }
            displayName: { type: string }
            avatarUrl: { type: string }
            verified: { type: boolean }
        likes: { type: integer }
        createdAt: { type: string, format: date-time }
        pinned: { type: boolean }
        creatorHearted: { type: boolean }
    WatchParty:
      type: object
      properties:
        roomId: { type: string }
        videoId: { type: string }
        hostUserId: { type: string }
        inviteUrl: { type: string }
        participants:
          type: array
          items: { $ref: '#/components/schemas/Participant' }
        positionSec: { type: integer }
        playing: { type: boolean }
    Participant:
      type: object
      properties:
        userId: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string }
        muted: { type: boolean }
        speaking: { type: boolean }
        joinedAt: { type: string, format: date-time }
...